kettle parse: moved into git folder
