+++ git/parse_kettle_source.py
@@ -7,14 +7,16 @@ Python script to parse Kettle files and validate it against set standards
 '''
 
 import re
+import sys
+import logging
+import argparse
 import xml.etree.ElementTree as ET
+from names_regex import DEFAULT_NAMES as DEFAULT_NAMES
 
 # --------------------------------------------------------------------------------------------------
 
-# TODO: Complete default name list
-_RE_DEFAULT_JOB_NAME = re.compile('^Job', re.IGNORECASE)
-_RE_DEFAULT_TRANS_NAME = re.compile('^Transformation', re.IGNORECASE)
-DEFAULT_NAMES = [_RE_DEFAULT_JOB_NAME, _RE_DEFAULT_TRANS_NAME]
+PARSER_LOGGER = logging.getLogger(__name__)
+logging.addLevelName(logging.ERROR, "\033[1;41m%s\033[1;0m" % logging.getLevelName(logging.ERROR))
 
 # --------------------------------------------------------------------------------------------------
 
@@ -25,11 +27,33 @@ def step_notes(root):
     If there are no notes, fail
     :param root:
     """
+    PARSER_LOGGER.info('Checking for notes')
     step_notes_passed = True
     notepads = root.findall('./notepads/notepad/note')
     if not notepads:
         step_notes_passed = False
     return step_notes_passed
+
+# --------------------------------------------------------------------------------------------------
+
+
+def evaluate_entries(entries):
+    """
+    If any name is the same as the default name return false
+    :param entries:
+    """
+    step_names_passed = True
+    evaluated = False
+    for entryname in entries:
+        PARSER_LOGGER.info('Checking entry {}'.format(entryname.text))
+        for default_name in DEFAULT_NAMES:
+            if re.findall(default_name, entryname.text):
+                step_names_passed = False
+                evaluated = True
+                break
+        if evaluated:
+            break
+    return step_names_passed
 
 # --------------------------------------------------------------------------------------------------
 
@@ -40,13 +64,31 @@ def step_names(root):
     check if they have non default names
     :param root:
     """
-    step_names_passed = True
-    entries = root.findall('./entries/entry/name')
-    for entryname in entries:
-        for default_name in DEFAULT_NAMES:
-            if re.findall(default_name, entryname.text):
-                step_names_passed = False
+    PARSER_LOGGER.info('Checking Step Names')
+    type = root.tag
+    if type == "job":
+        entries = root.findall('./entries/entry/name')
+    elif type == "transformation":
+        entries = root.findall('./step/name')
+
+    step_names_passed = evaluate_entries(entries)
+
     return step_names_passed
+
+# --------------------------------------------------------------------------------------------------
+
+
+def evaluate_descriptions(entries):
+    """
+    If no description text fail check
+    :param entries:
+    """
+    step_descriptions_passed = True
+    for entrydesc in entries:
+        if not entrydesc.text:
+            step_descriptions_passed = False
+
+    return step_descriptions_passed
 
 # --------------------------------------------------------------------------------------------------
 
@@ -54,36 +96,63 @@ def step_names(root):
 def step_descriptions(root):
     """
     Takes root element and finds descriptions
-    If no descriptions exit for each entry - return False
-    TODO:
-    method throws - Handle
-    __main__:11: FutureWarning: The behavior of this method will change in future versions.
-    Use specific 'len(elem)' or 'elem is not None' test instead.
+    If no descriptions exist for each entry - return False
     :param root:
     :returntype boolean:
     """
-    step_descriptions_passed = True
-    entries = root.findall('./entries/entry/description')
-    for entrydesc in entries:
-        if not entrydesc:
-            step_descriptions_passed = False
+    PARSER_LOGGER.info('Checking descriptions')
+    type = root.tag
+    if type == "job":
+        entries = root.findall('./entries/entry/description')
+    else:
+        entries = root.findall('./step/description')
+
+    step_descriptions_passed = evaluate_descriptions(entries)
     return step_descriptions_passed
 
 # --------------------------------------------------------------------------------------------------
 
 
 def kettle_evaluate(root):
-    step_names(root)
-    step_descriptions(root)
-    return
+    evaluation_passed = True
+    if step_names(root):
+        PARSER_LOGGER.info('Step names check passed')
+    else:
+        PARSER_LOGGER.error('Step names checks failed')
+    if step_descriptions(root):
+        PARSER_LOGGER.info('Step descriptions check passed')
+    else:
+        PARSER_LOGGER.error('Step descriptions check failed')
+
+    if step_notes(root):
+        PARSER_LOGGER.info('Notes check passed')
+    else:
+        PARSER_LOGGER.error('Notes check failed')
+    return evaluation_passed
 
 # --------------------------------------------------------------------------------------------------
 
 
+def parse_command_line_args():
+    parser = argparse.ArgumentParser(
+        description='Parse the Kettle file evaluate against kettle standards')
+    parser.add_argument('--filename', '-f', type=str, help='filename of the file to evaluate')
+
+    return parser.parse_args()
+
+# --------------------------------------------------------------------------------------------------
+
 if __name__ == "__main__":
-    """
-    Change filename to test
-    """
-    TREE = ET.parse('/Users/nnair/Downloads/copyforpm/form_d_companies.kjb')
+
+    PARSER_LOGGER = logging.getLogger()
+    PARSER_LOGGER.setLevel(logging.INFO)
+    CONSOLE_HANDLER = logging.StreamHandler(sys.stdout)
+    CONSOLE_HANDLER.setLevel(logging.INFO)
+    CONSOLE_HANDLER.setFormatter(logging.Formatter('%(asctime)s - %(levelname)s - %(message)s'))
+    PARSER_LOGGER.addHandler(CONSOLE_HANDLER)
+
+    CL_ARGS = parse_command_line_args()
+    FILENAME = CL_ARGS.filename
+    TREE = ET.parse(FILENAME)
     ROOT = TREE.getroot()
     kettle_evaluate(ROOT)
